Remove transitive dependency netty5-codec-http from reactor-netty5-core
Move HTTP proxy functionality from ProxyProvider to HttpClientProxyProvider.

Fixes #1953
