+++ reactor-netty5-core/build.gradle
@@ -58,7 +58,9 @@ dependencies {
 	compileOnly "com.google.code.findbugs:jsr305:$jsr305Version"
 
 	api "io.netty:netty5-handler:$nettyVersion"
-	api "io.netty.contrib:netty-handler-proxy:$nettyContribVersion"
+	api("io.netty.contrib:netty-handler-proxy:$nettyContribVersion") {
+		exclude module: "netty5-codec-http"
+	}
 	api "io.netty:netty5-resolver-dns:$nettyVersion"
 	if (!"$nettyVersion".endsWithAny("SNAPSHOT")) {
 		if (osdetector.classifier == "osx-x86_64" || osdetector.classifier == "osx-aarch_64") {
