+++ reactor-netty5-http/src/main/java/reactor/netty5/http/client/HttpClientProxyProvider.java
@@ -19,8 +19,10 @@ import io.netty.contrib.handler.proxy.HttpProxyHandler;
 import io.netty.contrib.handler.proxy.ProxyHandler;
 import io.netty5.handler.codec.http.headers.HttpHeaders;
 import reactor.netty5.transport.ProxyProvider;
+import reactor.util.annotation.Nullable;
 
 import java.util.Objects;
+import java.util.Properties;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
@@ -42,6 +44,17 @@ public final class HttpClientProxyProvider extends ProxyProvider {
 		 */
 		T httpHeaders(Consumer<HttpHeaders> headers);
 	}
+
+	static final String HTTP_PROXY_HOST = "http.proxyHost";
+	static final String HTTP_PROXY_PORT = "http.proxyPort";
+	static final String HTTP_PROXY_USER = "http.proxyUser";
+	static final String HTTP_PROXY_PASSWORD = "http.proxyPassword";
+	static final String HTTPS_PROXY_HOST = "https.proxyHost";
+	static final String HTTPS_PROXY_PORT = "https.proxyPort";
+	static final String HTTPS_PROXY_USER = "https.proxyUser";
+	static final String HTTPS_PROXY_PASSWORD = "https.proxyPassword";
+	static final String HTTP_NON_PROXY_HOSTS = "http.nonProxyHosts";
+	static final String DEFAULT_NON_PROXY_HOSTS = "localhost|127.*|[::1]";
 
 	final Supplier<? extends HttpHeaders> httpHeaders;
 
@@ -81,6 +94,67 @@ public final class HttpClientProxyProvider extends ProxyProvider {
 	@Override
 	public int hashCode() {
 		return Objects.hash(super.hashCode(), httpHeaders.get());
+	}
+
+	@Nullable
+	static HttpClientProxyProvider createFrom(Properties properties) {
+		Objects.requireNonNull(properties, "properties");
+
+		if (properties.containsKey(HTTP_PROXY_HOST) || properties.containsKey(HTTPS_PROXY_HOST)) {
+			return createHttpProxyFrom(properties);
+		}
+
+		return null;
+	}
+
+	/*
+		assumes properties has either http.proxyHost or https.proxyHost
+	 */
+	static HttpClientProxyProvider createHttpProxyFrom(Properties properties) {
+		String hostProperty;
+		String portProperty;
+		String userProperty;
+		String passwordProperty;
+		String defaultPort;
+		if (properties.containsKey(HTTPS_PROXY_HOST)) {
+			hostProperty = HTTPS_PROXY_HOST;
+			portProperty = HTTPS_PROXY_PORT;
+			userProperty = HTTPS_PROXY_USER;
+			passwordProperty = HTTPS_PROXY_PASSWORD;
+			defaultPort = "443";
+		}
+		else {
+			hostProperty = HTTP_PROXY_HOST;
+			portProperty = HTTP_PROXY_PORT;
+			userProperty = HTTP_PROXY_USER;
+			passwordProperty = HTTP_PROXY_PASSWORD;
+			defaultPort = "80";
+		}
+
+		String hostname = Objects.requireNonNull(properties.getProperty(hostProperty), hostProperty);
+		int port = parsePort(properties.getProperty(portProperty, defaultPort), portProperty);
+
+		String nonProxyHosts = properties.getProperty(HTTP_NON_PROXY_HOSTS, DEFAULT_NON_PROXY_HOSTS);
+		RegexShouldProxyPredicate transformedNonProxyHosts = RegexShouldProxyPredicate.fromWildcardedPattern(nonProxyHosts);
+
+		Build proxy = new Build()
+				.type(Proxy.HTTP)
+				.host(hostname)
+				.port(port)
+				.nonProxyHostsPredicate(transformedNonProxyHosts);
+
+		if (properties.containsKey(userProperty)) {
+			proxy = proxy.username(properties.getProperty(userProperty));
+
+			if (properties.containsKey(passwordProperty)) {
+				proxy = proxy.password(u -> properties.getProperty(passwordProperty));
+			}
+			else {
+				throw new NullPointerException("Proxy username is set via '" + userProperty + "', but '" + passwordProperty + "' is not set.");
+			}
+		}
+
+		return proxy.build();
 	}
 
 	static final class Build extends AbstractBuild<Build> implements Builder<Build> {
